Separating optional and obligatory rules
The bug found in the previous commits is due to an operation that, in hindsight, we probably don't want to do anyway.  This commit rethinks the way we express insertion rules and optionality.

Stepping back, we want to allow rules with empty/nullable "from" fields, in order to express processes like unconditional insertion.

   ε -> b

Just putting this into our existing rules doesn't work.  A rule like `a -> b` puts a `Match(NotContains(a))` in between the relevant material, which makes the application of the rule obligatory.  This isn't going to work for a `ε -> b`, Match(NotContains(ε)) would be empty.  (By the meanings of our nots and contains, this is empty -- every string "contains" ε in our sense of contains.)  To get this working right, we just insert a `Match(.*)` instead.

What I didn't really appreciate, but which is obvious in hindsight, is that this is just the formula for optionality.  And `ε -> b` is indeed an optional rule, it doesn't happen *every* time there's an ε -- any time you look for an ε you'll find one.  This means "you're allowed to insert a `b` anywhere" not ``you must insert a `b` everywhere".

So, what we were really doing is automatically switching to an optional rule semantics whenever the "from" material was nullable.  But do we really want to do that?  It can lead to surprises; consider the following two rules:

   * a -> b
   * ε|a -> b

These actually have very different behavior with respect to `a`; the first one replaces *every* `a` with `b`, whereas the second results in any subset of the `a`s in the sentence replaced with `b`.

We should probably make the programmer be explicit about what they want w.r.t. optionality/obligation, rather than switching the semantics behind the scenes.  This makes it a little harder for the programmer -- they'll have to understand optionality in order to program an unconditional insertion rule -- but I'm okay with making it harder to write unconditional insertions, those are dangerous anyway.

So this commit adds an optionality parameter to ReplaceGrammar, and uses that to decide whether the rule is optional, rather than deciding based on the delta of the from material.  I've reworked the tests that rely on this to use the new parameter, but I haven't yet hooked it up to the sheets interface.  We'll need to discuss in a meeting the syntax we want for optional rules.