Reworking grammar length estimation
In the previous commit, I differentiated between optional and obligtory rules, and made it so this choice is no longer automatic.  If the programmer wants an unconditional insertion rule, it has to be optional (or else it won't work).  But now that it's the programmer's responsibility, we should really be warning them when they choose the wrong one.

But that's actually a bit tricky!  We can't just do the same thing as before (test for nullability during RepeatGrammar.constructExpr), that's what got us into this trouble in the first place.

For this warning, though, we don't have to get it 100% right, nothing goes too awry if we're wrong, so we can do it earlier.

This is, conceptually, very much like Grammar.potentiallyInfinite: that's looking for the maximum possible length of an output, this is looking for the minimum.  So this commit reworks Grammar.potentiallyInfinte into Grammar.estimateLength.  This is also a good opportunity to make this function sensitive to tapes, which we've been meaning to do.

I've temporily shut off our safety-wrapping of grammars in Count.  That was fairly crude, and we've since come to a consensus that Count itself is to be avoided anyway.  Rather than figure out how to connect this new, more fine-grained system to our crude Count wrapping, I'd rather redo that system to something that's likewise more fine-grained, using CountTape instead.