Making infinity protection use CountTape
In practice, most clients don't want to generate forever from an infinite grammar, so we let clients cap the size of outputs, at least for infinite grammars.  We used to go through the grammar with Grammar.potentiallyInfinite(), and if it was potentially infinite, wrapped the whole thing in a CountGrammar.  But this wasn't tape-sensitive, it happened if *any* tape was infinite, and capped the total number of characters across all tapes.

In the last commit, I changed this to Grammar.estimateLength(), which *is* tape-sensitive.  So in this commit, I rework the infinity protection so that it only applies to tapes that can actually be infinite.

Since it was getting a little complex, and is used in two places (interface.ts and executeTests.ts), I factored it out into its own function.