Simplifying Count
While working on the counting logic, this was a good opportunity to make a simplification.

Last month (or so), we made a change so that Count (then: CountTape) allowed infinite generation on tapes that it didn't care about.  (E.g. if X defined t1/t2/t3, Count({t1:3,t2:5}, X) would be the same as saying Count({t1:3,t2:5,t3:Infinity}, X).)  Because they allow unrelated material through unhindered, this meant that now Counts could be nested.

	(1) g = Count({t1:3,t2:5}, g)

	(2) g = Count({t1:3}, g)
       g = Count({t2:5}, g)

Since much of the complexity of Counting dealt with maintaining its tape dictionary, and it's always possible to do decompose a Count like this, this offers an opportunity to make a substantially simpler CountGrammar/CountExpr, that only cares about one tape, and nest them when multiple are needed.  This commit does that.

It's still possible to do (1), with a multi-tape dictionary, but now rather than the Count() convenience function constructing a complex CountGrammar, Count() iterates through that dictionary and makes a nested structure of simple CountGrammars.  The downside is that we lose the shorthand to just say Count(100, X) and have that 100 apply to every known tape; we have to be explicit now.  (Tape information isn't available at the point the convenience function is invoked.)

Also made some more Count tests, and fixed a bug: when short-circuiting due to a result being too long, we were `return`ing rather than `continue`ing.  If there was a series of results of different lengths (e.g. due to atomic tokens or epsilon returns), as soon as one result is too long, we discard the rest even if they'd be valid.  CountTest 8c illustrates this: it fails with `continue` and succeeds with `return`.