Problematic rule checking and tests
(1) A few commits ago, we made a change to rules, that instead of nullable-from rules automatically triggering optionality, the programmer had to express it as an optional rule.  (How, exactly, is TBD.)  Given this change, we should warn if the programmer makes a nullable-from obligatory rule.  The commits since then were originally motivated by a need for some sort of nullability-check that wasn't itself delta() (because doing so raises a bigger issue) and could be called at will, we now have that in estimateLength().

This commit checks for this situation and signals a warning if it's not true.  (A warning, not an error, since an obligatory unconditional execution is safely executable, it's just probably not what the programmer intended.)

(2) We discussed having "to" tapes just be plaintext rather than regexes.  While I can there being a valid use of alternation in "to", I can't see a use of "to" regexes that's valuable enough to make up for the dangers, so I've turned this off.

In the meantime, I've also made it so that we error and epsilon when we have an infinite toGrammar, just in case.  While the programmer can't currently construct one of these in the sheets syntax, they're still semantically possible in (say) tests, so I added a check for safety.

(3) These sanity checks happen in their own pass, right before joining them to the grammar.  Meanwhile, the identity of this next pass was getting confusing (e.g. we didn't use replaceRules at all, and only used replaceRulesOld).  I removed replaceRules, and renamed replaceRulesOld as constructRuleJoins.

(4) I had left in a spurious epsilon alternation in ReplaceGrammar.constructExpr from prior experimentation, I took it out here.

(5) The ruleFromRegex etc. tests weren't necessarily testing *regex* syntax, because they were just alternations, and alternations are valid in plaintext now.  I renamed these ruleFromAlt for clarity, and added ruleFromRep/etc. to test a repetition instead.

(6) The new test rule/rulePreRep illustrates a newfound bug, that the rule `b -> v | da+_` fires correctly for "daba" but doesn't fire for "daaba".  It feels like this is an issue with `Short`ing, maybe the scope of a short is too narrow around the preContext whereas it should be around the whole from material.