Fixing "bugged" rule test
Fwhew, the apparent bug discovered in the past commit was just a mistake in the test itself -- I got the regex wrong.

I expressed the rule as

   b -> v | da+_

But the scope of the + was wrong, in Gramble it scopes over the whole (da), meaning that this matches "dab" and "dadab".

I should have expressed the rule as follows:

  b -> v | d(a)+

This matches "dab" and "daab" as I had intended.