Rule context parsing
(1) We can now parse rule contexts (e.g. `#_b` meaning "at the start of the word and before `b`) into an object called a RuleContextGrammar, which contains preContext/postContext content as well as begin and end flags.  This object isn't yet hooked up to the rest of the rule system, though, for now all we can do is parse the cells.

(2) Doing this also revealed a bug in the miniParser, escaped operators like `\|` could in some contexts be interpreted as operators.

The reason was because we were using backslashes as guides to tokenization alone, and unescaping after tokenization.  This meant that, often, a sequence like `\|` causes the `|` to attach to the previous token and thus not be recognizeable as an operator anymore.

But when `\|` doesn't have anything to attach to because it's the first thing in the string, or the previous thing can't be interpreted as an Unreserved, it becomes its own token, then gets unescaped, then gets parsed as an operator.

This is fixed by moving the unescaping step to MPUnreserved instead; that way it's still `\|` throughout the parsing process and doesn't get misidentified as `|`, and the backslash is only removed right before creating the LiteralGrammar.