Making replacement rules use contexts
This commit hooks up rule contexts (e.g. `#_b`) to the rest of the system.  This hit a bunch of small snags because rule contexts aren't much like other things in the system.

(1) Empty contexts are valid but we couldn't parse them, we needed an MPEmpty grammar to handle not having a context at all.  (NB: This is like symbol parsing, and unlike plaintext/regex parsing.  Plaintext and regex are MPRepetitions at root because a sequence of plaintext/regex objects is itself a plaintext/regex.  But a sequence of symbols isn't a symbol, and likewise a sequence of contexts isn't a context.)

(2) When parsing and constructing headers, from/to/context became TapeNameHeaders surrounded by ParamNameHeaders.  The whole RuleContextGrammar ended up wrapped in a SingleTapeGrammar, rather than the individual pre and post children.  (Because at the point we're nesting these, the RuleContextGrammar doesn't exist yet.)  This isn't wrong, they do share a tape, but it becomes awkward later to grab RuleContextGrammar.preContext and .postContext when it's so deeply nested in LocationGrammars and SingleTapeGrammars.

This nesting was an artifact of old-style rules, however, and old-style rules no longer exist in the sheets syntax.  Now, it's easier to just create an atomic To/From/Context header and wait until headerToGrammar to worry about this stuff.  At that point, the RuleContextGrammar exists, so we can just wrap the pre/post children rather than the whole thing.

(3) CreateGrammars.handleHeaderContentPair was creating a redundant LocationGrammar.  Didn't really matter... until we were awkwardly trying to reach inside them to grab rule pre/posts.

(4) It was possible to have the same parameter expressed multiple times in a single grid, like

          replace text:, from, from, to
                         a, b, c

This was equivalent to `from:ab, to:c`.  This was potentially useful back we we didn't have symbol embedding, because it was the only way to express something that we'd now express as `from: q{Vowel}`.  But now we can just say that.

That was okay (albeit weird) for "from" and "to", because those take regexes, and sequences of regexes are regexes.  But as mentioned above, sequences of contexts aren't contexts.  This is the first time a named param *couldn't* be a sequence, actually, so we had to change up param handling to forbid a sequence of `context`s.  While I was at it, I also forbid sequences of `from`s like the above, as well as `to`/`pre`/`post`.

`unique` and the secret param `__` (for ordinary content) can still be sequences, though.  (After all, big things go wrong if ordinary content can't be joined into a sequence!)

(5) Made tests for simple rules using contexts.  They work!