Bug found in begin+end rules
I was adding some new sheets tests to make sure rule contexts were working right, and uncovered a bug in replacements.  I've also replicated this bug in testReplace.

The test is simple: `a -> x | #_#` will erroneously exclude any input that starts with `a` other than `a`.  That is:

   `a` becomes `x` (correct)
   `b` becomes `b` (correct)
   `ba` becomes `ba` (correct)
   `ab` fails (incorrect)

It turns out that all the tests for begin=true, end=true also required minReps=1.  That means we were missing the erroneous case `ab` -> null, since that was also excluded by minReps=1.