Rescoping repetitions in regexes
Previously, all operators in regexes scoped over the longest possible sequence of literal characters, e.g.:

   * abc|def == (abc)|(def)
   * ~abc == ~(abc)
   * abc+ == (abc)+
   * abc* == (abc)*

But I found myself making errors about repetition:

   * this is not how regexes are parsed in most languages

   * in the Gramble context where we most often use repetition operators, rule contexts, we're typically thinking at the character level.  E.g., how often do we need to say things like "d(a)+" vs. "(da)+"; the latter is rather odd as a context for a rule!

So we decided in the last meeting that we should scope repetition operators narrowly, only around the previous unit.

This was a bit tricky because character-level parsing is quite different from what we were doing before, we typically tokenize letter sequences together because they often act as units otherwise.  One option is just to give up on doing this, tokenize at the character level, and reassemble things like symbol names when necessary; the other is to make an Unreserved parser that takes the longer string and just chomps off one character.  I did the latter here.

There's one more complication; we end up with MPRepetitions at two levels of the grammar, and the lower-level one can occur unparenthesized in the top-level one.  This results in ambiguity, so I eliminated the top-level sequence.  This ends up changing the scope of the weird examples like `foo moo | goo`, this used to mean `foogoo | moogoo` because concatenation outscoped alternation, but now alternation outscopes concatenation so it's `foomoo | goo`.  I don't like this as much, but in the end I think it's easier than having to explain that concatenation occurs twice in the precedence hierarchy.