Interpreter API now qualifies names
Yesterday I was trying to make some changes to how Default-ing works (the way that bare references like "Verb.Intransitive" have particular interpretations), and got stymied, because there are still some odd bits about how that works.

I realized one of the causes for this difficulty, which is that Interpreter didn't actually know how to resolve symbol names.  From Interpreter's point of view these were just strings.  If a symbol has two possible ways of referring to it, like "Verbs.Intransitive" and "Verbs.Intransitive.Default", Interpreter expects both of those to be in the symbol table.  It didn't have any functionality that let it understand that those are the same thing, it just looks up the string and if it's not in there, that's an error.

We *could* just reduce those names to a canonical form in Interpreter, like just cutting out every instance of ".Default", but in the long run this would still be brittle.  The deeper issue is that Interpreter and the Grammar name qualification pass have *different* ways of understanding names, and any time we want to make any changes to the latter, the former will break.

So I think it's better in the long run to make it so that these share the exact same understanding of names -- that Interpreter resolves a reference to a symbol using the same function that we use in QualifyNames.

We didn't do this before because Interpreter didn't, at the point of calling (say) .generate(), actually have access to the structure that would let it qualify names correctly.  That structure is already flattened at that point, it's a flat CollectionGrammar of fully-qualified names.

So now Interpreter keeps around a copy of the grammar from right before we flatten it, for whenever it needs to qualify a name.  (This isn't that strange, we already keep around another previous stage, the workbook stage, which we need to do source-to-source transformations like single-paging.)

Anyway, the big picture here is that if/when the nature of names/defaults/etc. changes in the future, Interpreter's treatment of names won't get out of sync, because it'll be using the same resolution code.