Finer-grained name resolution errors
We're working towards the idea that not every symbol in the grammar actually has a default interpretation (because auto-defaulting has results that are sometimes surprising to the programmer).

This means that there will be a subtle but important difference between two error cases:

  (1) The symbol the programmer invokes can't be found at all.
  (2) The symbol the programmer invokes refers to a Collection, and that collection doesn't have a default symbol.

These are important to distinguish to the programmer because the latter may be surprising.  From the programmer's point of view, the symbol in (2) *exists*, it refers to a thing that they programmed, so it's an odd error to receive that the symbol can't be *found*.

Also, it's important during our internal search to distinguish between not-finding from finding-something-erroneous.  When we haven't found it AT ALL in a given scope, the right thing to do is continue the search in the enclosing scope.  When we've found it and something's wrong, we shouldn't search for another one of the same name in enclosing scopes.  That would be surprising to the programmer, if an error with a local X caused X to refer to something different!