Removing abbreviated names from symbol tables
There was a longstanding duplication in our symbol tables, done out of convenience but clearly a bad idea even at the time.  Default symbols like "Verbs.Intransitive.Default" have two ways of referring to them, that and "Verbs.Intransitive", and both of those existed in the symbol table, referring to the same grammar.  (Or, to be precise, referring at one point to the same grammar object, which eventually gets cloned into different, albeit equivalent, descendants.)

We could have, and should have, resolved the latter name to the former, or the former name to the latter, and gotten rid of the other one in the symbol table.  There was one major barrier to this, that Interpreter was stupid about symbol names, it wouldn't know if a client asked to generate from "Verbs.Intransitive" that they really meant "Verbs.Intransitive.Default".  We had to keep both names in the symbol table so Interpreter could ask for either of them without throwing a symbol-not-found exception.  As of recently, this has been addressed; Interpreter nows resolves symbol names by the same algorithm as we use elsewhere, and so whatever we choose to do about resolution, Interpreter won't get confused.

So this commit follows through with that:

(1) Name resolution is now always to the full form including any "Defaults".

(2) I removed the code that auto-assigned default symbols to abbreviated names.  So now, abbreviated names like "Verbs.Intransitive" don't exist in the symbol table at all.

(3) A tricky consequence arose in unit test execution.  CollectionGrammars begin with a default symbol of "", and in normal execution we replace that with whatever symbol the client requests.  The corresponding CollectionExpr is basically just a wrapper around the referent of that symbol that also pushes the symbol table into the env.  If we don't happen to find the symbol, the referent is just an epsilon.

When executing unit tests, we don't replace the default symbol, it just stays as "".  We no longer find that (because it no longer exists), so the referent is epsilon.  However, passing in an epsilon to constructCollection() doesn't give us a CollectionExpr, it gives us an epsilon.  (If it didn't, CollectionExprs wouldn't simplify by the end.)  In becoming epsilon, it loses the symbol table, and we need that symbol table to run the unit tests.

I changed the CollectionGrammar default symbol to "Default" (which it should now be anyway) but this doesn't really address the bigger issue.  At the moment `Default` is sure to exist, but soon that won't be true.

For now, I've made it so that, when we don't find the default, we construct a CollectionExpr directly without the use of constructCollection(), so that it doesn't prematurely simplify.  This isn't a general solution, it's janky but ultimately the whole thing we're doing to construct unit test execution environments is janky, I'm not sure there's a non-janky solution until we solve the bigger problem.

(4) Getting rid of abbreviated names in the symbol table unearthed another piece of code that wasn't resolving names properly, testHasTapes.  Now, testHasTapes resolves the name through Interpreter, rather than trying to look it up directly in the grammar.