Small fix to ErrorHeader
ErrorHeader was inheriting its mapChildren from its ancestor, tapeNameHeader.  Nothing should inherit mapChildren, that always has to be overridden (or else the child class actually becomes the parent class!).  We don't ever happen to call this, so it didn't cause any current bugs, but it could have been a nasty surprise in the future.

In any case ErrorHeader should not have been a descendant of TapeNameHeader in the first place, that was laziness.  Now it's just a direct descendant of Header like it should be, with its own mapChildren.