Removing all defaulting
As mentioned several commits ago, the purpose of this branch was to experiment with disentangling defaulting behavior with the ability to refer to any element in a collection (now expressed as e.g. `Verb.All` instead).  The conflation between the two was getting surprising and confusing.

This took several commits to pave a smooth path for this change, but I basically managed it...

... and have come to the conclusion that the new defaulting behavior (in which some symbols have default interpretations and others have none) is equally surprising and confusing.  It's rather easy to end up in situations where there are defaults (named or auto) that refer to things that don't have interpretations, and hard to reason about why.

So I got to thinking, why do we need this behavior?  Who's using it?  It's not widely used in the sheets themselves; in the past programmers have said they prefer to be explicit.  One programmer does use it, but only to get the "All" behavior, which now has a different name, `X.All`.

There's really only one common place where defaulting saves time/thought/typing, and that's in sheets tests.  It lets us express the test without specifying exactly what symbol to generate from.  That's handy but testing alone doesn't justify having such a complex and hard-to-reason-about system in the language itself.

So I decided to axe defaulting.  Making any serious change to defaulting will require writing 100+ sheets tests anyway, so we might as well rewrite them all to be explicit about what symbol to generate from.  (Or, rather, generate a reasonable default symbol using convenience functions in the test utils.)

Sheets tests now use a convenience function testProject that takes unordered keyword params.  (I didn't realize how to do this in TypeScript until just now.)  Unordered keyword params are handy for these tests, that way we don't have to fill out DEFAULT, DEFAULT, DEFAULT for positional params in between the initial params and whatever later param we actually want to test.

This requires rewriting every test, so I'm only about half done as of this commit.