Catching missing files/symbols in tests
Formerly, if we referred to a non-existent source file, or a non-existent symbol in a grammar, we threw an exception and didn't catch it within the tests.  This led to the situation where if there was even a single missing file/symbol, we couldn't run (or even collect) any other tests until this was addressed.

This commit wraps the dangerous queries in try/catch, and if the throw causes there to be a special (and always-failing) test, so that it's brought to our attention but still lets the test suite move on to the next test.

Also fixed up testSheetCell and testEmbed to reflect the new lack of defaulting behavior.