More test improvements
(1) Mocha Explorer wasn't finding the parse tests anymore, so I added that directory to the config.

(2) Adapted the error/hide/rename sheets tests

(3) Made it possible to specify a filename in testProject rather than always have it built.  That's necessary when we're testing things about sheet naming (e.g. that one can't name a sheet a reserved word).

(4) On that note, something about our recent changes restored the ability to name sheets to reserved words.  I added a test for this in parseSheets.