Finishing adapting tests
This adapts the "collection", "multi", "filter", "rule", and "test" sheets tests to the new defaultless semantics, and the new sheet rule testing regime (in which we use a test({}) function with optional params).

(1) The "collection" tests obviously had to change a lot -- that's what defaults apply to, after all -- and many were removed because they were testing behavior that no longer exists.  A few tests in "multi" had the same (since "multi" is testing collections at the project level).

(2) We're going to remove `pre` and `post` as valid headers, so while I was at it I removed/adapted the relevant tests in "rule" to use `context` instead.