Small sheet rule bugfixes
Switching from pre/post rules to context rules unearthed a few bugs:

(1) We weren't correctly localizing the error when a multi-tape symbol was embedded in context.  The problem was when creating ReplaceGrammars, when we're unpacking RuleContextGrammars to get their pre/post arguments.  If the RuleContextGrammar is wrapped in a LocatorGrammar (which it is, here), we unwrap it, but instead of doing anything with that position, we just throw it out.

Now we use that position to re-wrap the RuleContextGrammar's pre and post children before constructing the rule.

(2) If rule sanity checking caused *all* of the rules in a JoinRule to become invalid (which we indicate by replacing them with an EpsilonGrammar), we weren't getting the right answer (which is to do nothing to the input grammar).  Now we're checking for that and returning early.