Refactoring erroneous rule handling
The previous commit involved a bit of an awkward solution -- letting JoinRuleGrammar have epsilon rules and then filtering those out during the rule-joining process.  To be completely typed, this meant that .rules had to be (ReplaceGrammar|EpsilonGrammar)[], making us do various type checks/narrowing when we need to be sure.

Actually, we could still call this ReplaceGrammar[]; TypeScript can't catch the falsehood here for some reason.

Regardless, there's awkwardness here, and it's really because we're handling this problem in the wrong place.  EpsilonGrammar rules don't *mean* anything, they're just a dummy value used when a rule has errored so bad it ceases to exist.  Even though they're a possible result of ReplaceGrammar.mapChildren, it doesn't mean we should be automatically allowing those in JoinRuleGrammar.rules.

So now, JoinRuleGrammar.mapChildren filters out dummy rules, ensuring that the type of .rules really is ReplaceGrammar[].

Meanwhile, we also check in SanityCheckRules that there do exist some rules in JoinRuleGrammars.  If there aren't, we warn the programmer and eliminate the JoinRuleGrammar entirely (that is, replace it with its child).

We could have done this in JoinRuleGrammar.mapChildren too, but it felt strange to fire a programmer error/warning inside mapChildren.  That's a very generic place to try to construct a meaningful error message; mapChildren is called constantly for a variety of reasons.  So I prefer to keep that functionality in a pass.