Using Partial<T> for sheets tests
I just learned about Partial<T> and it's a bit cleaner than what we were doing with ProjectTest.  It's a built-in template that you apply to a record type to make every member optional.

The types and the generated code end up being identical, it's just that it helps you distinguish between the ideas:

    * This is a record type with an `.a` that's a `string|undefined` and  a `.b` that's `string|undefined`...

    * This is a record type with an `.a` that's a `string` and a `.b` that's `string`... but we let you construct this record only partially (e.g. because we're going to set defaults later).