Universal mapChildren
I've been playing with Components/Passes and the related machinery to try to make them easier to write (e.g., so that someone else can contribute a new Component/Pass, without necessarily being fluent in the Result monad).  A lot of this hasn't been fully successful.

But this one notched a win: I managed to make one version of Component.mapChildren that works for (almost) everything.  This was the main barrier to someone else making their own Tst or Grammar, that you have to be able to write a mapChildren method, and those require fluency in Results.

In order to do this, I had to give up on the code inside mapChildren being type-checked.  There's no type-safe way to iterate through the heterogeneous properties of a given class instance.

From the outside, though, it's actually safer than what we've been doing.  As a recent commit noted, if mapChildren was inherited from an ancestor, that causes bugs; same thing if you accidentally make an X.mapChildren that returns a new Y.  This can't do make that mistake, because it clones the object programmatically.

Some snags:

(1) This causes bugs if it copies ephermeral state like ._tapes or .fromTapeName.  This is a bit of a hack, but I solved this by making it consistent through the Component hierarchy that all and only ephemeral members started with "_", and then skipping those when cloning.

Also, all these _ members have to allow undefined, it doesn't work if they have default values like .fromTapeName did.  (If there's a way to grab those default values, I don't know what it is.)

(2) CollectionGrammar and LocatorGrammar need to do some setup/teardown before calling super.mapChildren.

(3) TstOp isn't working with this yet and I think it might be because Ops have all those get/readonly properties.