Removing vestigial old-replace code
Old-style replaces are no longer possible to express in sheets, but there were still a bunch of objects/handlers for them, in parallel to the new-style replaces.  In the relevant part of the codebase, the former were always named with `Replace`, the latter with `ReplaceTape`.

This eliminates the unused `Replace` code, and renames `ReplaceTapeX` stuff as `ReplaceX`.