Value-space pass + unconditional insertion bugfix
(1) This commit experiments with a kind of Pass that doesn't require the programmer to use/understand the Result system.  They're passes where all the code we actually write is in "value space" (we're acting as if functions take and return Components) as opposed to "Result space" (where we're juggling Result monads and in order to do anything we have to bind into them).

The Pass as a whole still has the same type, it's just that instead of overloading the transform function (and handling mapping children yourself and returning a result) you overload a Component => Component function postTransform.  We handle all the mapping of children beforehand (this is the most common order of transformation), and also wrap the call of postTransform in a try/catch block that catches any thrown errors and integrates them with our Result handling.

Exactly what the programmer should throw, I haven't quite decided.  In order to be integrated with Result handling, we basically need a message and a repair (the thing that we should put in the AST given that we failed to actually transform the thing we wanted to).  In a lot of cases the repair is just TstEmpty or EpsilonGrammar, though, meaning quite a lot of times the programmer could just throw a message and we do a default repair behind the scenes.  At minimum, the programmer could just throw a single string and we construct a Result out of it as best we can.  In any case, I haven't settled on any one best solution.

So far, this is working a little better than things like `g.err(msg).bind(g => g.child)`.  That was always a little awkward, because you couldn't just say `g.child.err(msg)` (that localizes the error to the child, which is often the wrong place).  It's rare that we want to localize the error to the repair.  In our error cleanup code, we're localizing that (when possible) to the original grammar -- the one passed into `.transform` in the first place -- which is usually where we want it to go.

(2) While I was re-doing the errors in SanityCheckRules, I realized we got one wrong.  We were identifying some rules wrongly as unconditional insertions: specifically, ones with empty pre/from/post but that had `.beginsWith` or `.endsWith` being true.  These are conditional, and don't pose any problems, so we shouldn't have been warning/erroring about them.  I added some tests to testSheetRule to ensure we're distinguishing real unconditional insertions from conditional insertions.