Adapting grammars to discriminated unions
I converted Grammars to be a discriminated union like Headers, and the passes to use g.tag rather than g.constructor.  It worked pretty well -- between this and the new recursion strategies (letting the user pretend they're working with values), passes are becoming a lot simpler-looking.

Even aside from the discriminated union bit, it turns out it's handy in TypeScript to have a `public readonly tag = "whatever"` member in every class.  Doing so catches a lot of type errors that you'd otherwise miss because of TypeScript's structural typing.

Like, if you are, say, writing a method `.handleStarts(g as StartsGrammar)`, and you need to write a similar one for Ends and make a copy-paste error to get `.handleEnds(g as StartsGrammar)`, this succeeds!  It's not actually erroneous to pass in an EndsGrammar to the latter as an argument, because the two classes are structurally equivalent.  TypeScript can't tell the difference.

Now this one isn't a bug exactly, because all those types are erased anyway, but it's sloppiness, I didn't *intend* to write that.  I was surprised how much of that was found by adding a readonly `.tag`.