Fixing mapChildren for Ops and TSTs
(1a) I had had some trouble with auto-mapping children for TstOp.  I assumed this was because Ops themselves weren't automapping correctly, and this was correct.

It wasn't that the Op members were readonly, but there was something up with readonly members that were sets.  I'm still not certain *why* that didn't work, but in any case those don't have to be members anyway.  These would be better as pattern-matching functions.

(1b) So, I adapted Ops to use pattern-matching and discriminated unions instead.  It's much clearer now (we don't to jump around working out the inheritance to know what the, say, allowed params for TestNotOp is, it's easy to read off that function).

(2a) In doing this, I realized that the Tst objects had members that were neither Components nor Dict<>s (that is, plain {string:T} objects), but had classes.  CellPos is a key one.  For Components, we were cloning using Object.create, but not for anything else, and that was causing us to lose behavior like CellPos.toString().  Now we clone all objects just in case.

(2b) We can't just use the exact same map function for the two, though, because Components we don't copy over keys starting with "_".  We can't do that with a generic objects, they're totally allowed to have keys that start with "_"!  TstParams.params often contains that because the default blank param was "__".  CollectionGrammar.symbols has keys that are used-generated symbols and those can start with "_".

(2c) While I was at it, I realized the default param shouldn't be "__" -- it should be something that's not a valid identifier, just in case we allow user-generated param names in the future.  I changed it to "$plain".  I also realized this was a magic constant throughout the code, now it's PLAIN_PARAM.

(3) I moved towards an auto-ID system for ops where we have a consistent ID schema based on its structural shape (e.g. always starting with the `.tag`, which we didn't always do before, recursing to the child every time there's a `.child`, etc.).  It should be rich enough to be parsed back into the original if we ever need to do that.  I chose s-exprs as the output format because those are easy to parse and unambiguous.

Then I did the same with headers.