miniParser now "value-space" recursion strategy
Originally, miniParser was just an ordinary function from values to values; it didn't use the Result monad.  You had to make an ErrorHeader or ErrorOp etc., and then to collect the messages in them you had to do some sort of recursive pass through the resulting tree.

Then I converted it all to use Results, which means that errors/warnings accumulate in the background.  However, a ton of otherwise-pretty-trivial construction functions become hairy to write.

This is basically the same problem as we encountered with mapChildren and Passes, and I thought the same solution would work here.  It did.

   * It used to be that all the functions were (abstracting away a bit) `T1 => Result<T2>`, both the big parse function as a whole and the little adapter functions (here, the `constr` arguments) that we pass into it.

   * The big parse functions are still `T1 => Result<T2>`, but the construction functions you pass in are now just `T1 => T2`.

   * If you need to err/warn in those functions, throw a Result.  We now always wrap the construction functions in a try block to catch Results, and accumulate the errors in the appropriate parsing functions.

The resulting code is much simpler.  (A little trickier in miniParser itself, but much simpler for anything that calls that.)

I also got rid of a little thing I noticed a while ago, that we were never using the text members of comment or error components, so I removed them.