Reconsolidating error system
Since you can throw anything, and the types of throws aren't part of the type signature of the function, that meant that we could throw a wide variety of error-y objects from Results to special error objects like PassError and GrammarError, to plain ol' strings.  This also allows for auto-repairing (like throwing an error without an explicit repair, and having the catcher supply a reasonable default).

The variety of new error types and possible defaults started to make this tricky, however, and after playing around with this the last few days, I've decided it's better to have one Right Way to Do It.

So now, if you want to throw an error in a Pass or miniParser or anywhere, throw a Result.

I had been avoiding that, because I got into my head that we didn't want a contributor to have to know about or use Results, but that's not really true.  Result is a totally reasonable error object with a lot of convenience features, there's no point in replicating all those again.  The real goal is that a contributor shouldn't have to understand the whole Result system to write basic code.  A contributor should just know "If you detect an error, throw this special error object with a proposed repair and a message".

I'm also working on re-working or eliminating the `Component.msg/err/warn` convenience functions.  These have caused a lot of trouble behind the scenes -- they wreak havoc with the type-checking, and their auto-localizing often puts things in the wrong place.  We had to have special versions in different parts of the Component hierarchy so that they return the right kinds of Results.  Anyway, I took them out of Ops and Headers and used some plain convenience functions `err(c, msg)` etc. instead, which don't localize and can more easily be typechecked.